Removes postgres middleware
