--- app/documentation/questions_redis/related.js
@@ -1,14 +1,0 @@
-const list = require("models/questions").list;
-
-module.exports = async function related(req, res, next) {
-  const path = require("url").parse(req.originalUrl).pathname;
-
-  if (["/questions", "/questions/ask"].indexOf(path) > -1) return next();
-
-  const related_tag = path.split("/").pop();
-
-  res.locals.related_tag = related_tag;
-  res.locals.related = await list({ tag: related_tag });
-
-  return next();
-};
